antes de agregar las validaciones a todos los modelos

+++ src/routes/registrarventas.js
@@ -9,12 +9,13 @@ const _ = require("lodash");
 router.post("/guardar", async (req, res) => {
   try {
     let ventasNuevo = new RegistroVentas(req.body);
-    ventasNuevo.usuario = req.user?.usuario ?? "usuariodesconocido";
-    let ventasAnterior = await RegistroVentas.findById(ventasNuevo._id);
-    if (ventasAnterior) { //si ya existe un registro anterior lo guarda en la parte de respaldos
-      await new RegistrosEliminados({ registro: ventasAnterior, borradoEl: new Date() }).save()
-      _.extend(ventasAnterior, _.pick(ventasNuevo, ["usuario", "ultimocambio", "tablas"]));
-      await ventasAnterior.save();
+    let usuario = req.user?.usuario ?? "usuariodesconocido"
+    ventasNuevo.usuario = usuario
+    let registroanterior = await RegistroVentas.findById(ventasNuevo._id);
+    if (registroanterior) { //si ya existe un registro anterior lo guarda en la parte de respaldos
+      await new RegistrosEliminados({ registro: registroanterior, borradoEl: new Date(), usuario, motivo: 0 }).save()
+      _.extend(registroanterior, _.pick(ventasNuevo, ["usuario", "ultimocambio", "tablas"]));
+      await registroanterior.save();
     }
     else
       await ventasNuevo.save();
@@ -36,12 +37,13 @@ router.post("/mover", async (req, res) => {
     let registroanterior = await RegistroVentas.findById(fecha) //si ya existe un registro anterior lo guarda en la parte de respaldos
     let { tablas } = registrorecuperado
     let ahora = new Date()
-    let nuevosDatos = { usuario: req.user?.usuario ?? "usuariodesconocido", ultimocambio: ahora, tablas }
+    let usuario = req.user?.usuario ?? "usuariodesconocido"
+    let nuevosDatos = { usuario, ultimocambio: ahora, tablas }
     if (registroanterior) {
       if (!sobreescribir)
         return res.send(registroanterior);
 
-      await new RegistrosEliminados({ registro: registroanterior, borradoEl: ahora }).save()
+      await new RegistrosEliminados({ registro: registroanterior, borradoEl: ahora, usuario, motivo: 1 }).save()
       await registroanterior.updateOne(nuevosDatos)
     } else
       await new RegistroVentas({ _id: fecha, ...nuevosDatos }).save()
@@ -64,10 +66,10 @@ router.route("/:id").get(async (req, res) => {
     let registro = await RegistroVentas.findById(fecha);
     let plantillas = await plantilla.find({}, { nombre: 1, _id: 0 }).sort("orden");
     let plantillaDefault = await plantilla.findOne({ esdefault: true });
-    let listaCamioneros = await Camioneros.findOne({}, { "camioneros.nombre": 1, _id: 0 });
-    let camioneros = listaCamioneros.camioneros.map(el => el.nombre);
+    let listaCamioneros = await Camioneros.findOne({}, { "camioneros.nombre": 1, _id: 0 })
+    let camioneros = listaCamioneros?.camioneros.map(el => el.nombre) || []
     let fechastr = fecha.toLocaleString(DateTime.DATE_HUGE);
-    res.render("registrarventas", { fecha: fecha.toFormat("d/M/y"), registro: registro, plantillas, plantillaDefault, fechastr, camioneros, DateTime });
+    res.render("registrarventas", { fecha: fecha.toFormat("d/M/y"), registro, plantillas, plantillaDefault, fechastr, camioneros, DateTime });
   } catch (e) {
     console.log(e)
     res.status(400).send("página inválida");
@@ -78,9 +80,12 @@ router.route("/:id").get(async (req, res) => {
     let datos = id.split("-");
     if (datos.length !== 3)
       return
-    let fecha = DateTime.fromFormat(id, "d-M-y");
-    let registro = await RegistroVentas.findById(fecha);
-    await new RegistrosEliminados({ registro, borradoEl: new Date() }).save()
+    let fecha = DateTime.fromFormat(id, "d-M-y")
+    let registro = await RegistroVentas.findById(fecha)
+    if (!registro)
+      return res.status(400).send("El registro no existe")
+    let usuario = req.user?.usuario ?? "usuariodesconocido"
+    await new RegistrosEliminados({ registro, borradoEl: new Date(), usuario, motivo: 3 }).save()
     registro.deleteOne();
     res.send("Exito");
   } catch (e) {
